Split ui into more components
