+++ app/Http/Controllers/HistoryController.php
@@ -16,15 +16,19 @@ class HistoryController extends Controller
     {
         $userId = Auth::user()->employee->id;
 
-        $timeRecords = TimeRecord::whereDate('recorded_at', now())
+        $currentMonth = today()->month;
+        $currentYear = today()->year;
+
+        $timeRecords = TimeRecord::whereMonth('recorded_at', $currentMonth)
+            ->whereYear('recorded_at', $currentYear)
             ->where('employee_id', $userId)
             ->orderBy('recorded_at', 'asc')
             ->get();
 
-        $timeRecordsResource = new TimeRecordByDayResource($timeRecords, today());
+        $timeRecordsThisMonth = new TimeRecordByMonthResource($timeRecords, today());
 
         return Inertia::render('History', [
-            'timeRecords' => $timeRecordsResource
+            'timeRecordsThisMonth' => $timeRecordsThisMonth
         ]);
     }
