+++ resources/js/Pages/History.vue
@@ -8,7 +8,21 @@
         </template>
 
         <PageContainer>
-            <HistoryCalendar />
+
+
+            <div>
+                <h2 class="h3">{{  displayedDate }} <small class="text-base">{{ displayedMonthYear }}</small></h2>
+                <div class="lg:grid lg:grid-cols-12 lg:gap-x-16">
+
+                    <!-- Calendar -->
+                    <MonthCalendar :today="today" :timeRecordsThisMonth="timeRecordsThisMonth.data" class="mt-10 text-center lg:col-start-8 lg:col-end-13 lg:row-start-1 lg:mt-9 xl:col-start-8"/>
+
+                    <!-- Timesheet section -->
+                    <TimeSheetSection class="mt-10 lg:col-span-7 xl:col-span-7" />
+
+                </div>
+            </div>
+
         </PageContainer>
     </AppLayout>
 </template>
@@ -16,10 +30,27 @@
 <script setup>
 import AppLayout from '@/Layouts/AppLayout.vue';
 import PageContainer from "@/Components/_util/PageContainer.vue";
-import HistoryCalendar from "@/Components/calendar/HistoryCalendar.vue";
+import MonthCalendar from "@/Components/calendar/MonthCalendar.vue";
+import TimeSheetSection from "@/Components/calendar/TimeSheetSection.vue";
+import {ChevronLeftIcon, ChevronRightIcon} from "@heroicons/vue/20/solid/index.js";
+import {computed, ref} from "vue";
 
 const props = defineProps({
-    timeRecords: Object,
+    timeRecordsThisMonth: Object,
 })
+
+const today = ref(new Date());
+const selectedDate = ref(today.value);
+
+const displayedDate = computed(() => {
+    const options = { weekday: 'short', day: 'numeric',  };
+    return selectedDate.value ? selectedDate.value.toLocaleDateString(undefined, options) : today.value.toLocaleDateString(undefined, options);
+});
+
+const displayedMonthYear = computed(() => {
+    const options = { month: 'short', year: 'numeric' };
+    return selectedDate.value ? selectedDate.value.toLocaleDateString(undefined, options) : today.value.toLocaleDateString(undefined, options);
+});
+
 
 </script>
